feat/AB#61579_DB-see-if-possible-to-have-different-instances-of-surveyjs fix: same id for dashboard form survey and dashboard filter survey fix: initialize survey structure of dashboard filter with no text feat: add a sticky close button to dashboard filter survey modal for a more user friendly experience feat: add setSurveyCreatorInstance is form.service to update the current survey creator instance with the needed question configuration refactor: add survey helpers to a new service to have a clean overview of the survey data/work flow in the code
