+++ libs/safe/src/lib/components/dashboard-filter/dashboard-filter.component.html
@@ -91,7 +91,7 @@
         <safe-button (click)="onEditFilter()">
           {{ 'components.application.dashboard.editFilter' | translate }}
         </safe-button>
-        <div #surveyCreatorContainer></div>
+        <div #dashboardSurveyCreatorContainer></div>
       </form>
     </div>
   </div>
