+++ libs/safe/src/lib/components/dashboard-filter/dashboard-filter.component.scss
@@ -69,5 +69,6 @@
 ::ng-deep .edit-filters-modal {
   mat-dialog-container {
     margin: auto;
+    padding-top: 0;
   }
 }
