+++ libs/safe/src/lib/components/dashboard-filter/filter-builder-modal/filter-builder.component.html
@@ -1,1 +1,12 @@
-<div class="w-full h-full" id="surveyCreatorContainer"></div>
+<div class="sticky top-0 z-10 bg-white">
+  <safe-button
+    class="ml-auto"
+    [isIcon]="true"
+    icon="close"
+    variant="danger"
+    [matTooltip]="'common.close' | translate"
+    (click)="cancelSurveyCreation()"
+  >
+  </safe-button>
+</div>
+<div class="w-full h-full" id="dashboardSurveyCreatorContainer"></div>
