+++ libs/safe/src/lib/survey/init.ts
@@ -34,6 +34,7 @@ import { initLocalization } from './localization';
  * @param authService custom auth service
  * @param environment injected environment
  * @param referenceDataService Reference data service
+ * @param containsCustomQuestions If survey contains custom questions or not
  */
 export const initCustomSurvey = (
   Survey: any,
@@ -43,19 +44,27 @@ export const initCustomSurvey = (
   formBuilder: UntypedFormBuilder,
   authService: SafeAuthService,
   environment: any,
-  referenceDataService: SafeReferenceDataService
+  referenceDataService: SafeReferenceDataService,
+  containsCustomQuestions: boolean
 ): void => {
-  // load widgets (aka custom questions)
-  SurveyJSWidgets.select2tagbox(Survey);
-  CommentWidget.init(Survey);
-  TextWidget.init(Survey, domService);
-  // load components (same as widgets, but with less configuration options)
-  ResourceComponent.init(Survey, domService, apollo, dialog, formBuilder);
-  ResourcesComponent.init(Survey, domService, apollo, dialog, formBuilder);
-  OwnerComponent.init(Survey, domService, apollo);
-  UsersComponent.init(Survey, domService, apollo);
-  // load global properties
-  ReferenceDataProperties.init(Survey, domService, referenceDataService);
+  // If the survey created does not contain custom questions, we destroy previously set custom questions if so
+  if (!containsCustomQuestions) {
+    Survey.CustomWidgetCollection.Instance.clear();
+    Survey.ComponentCollection.Instance.clear();
+  } else {
+    // load widgets (aka custom questions)
+    SurveyJSWidgets.select2tagbox(Survey);
+    CommentWidget.init(Survey);
+    TextWidget.init(Survey, domService);
+    // load components (same as widgets, but with less configuration options)
+    ResourceComponent.init(Survey, domService, apollo, dialog, formBuilder);
+    ResourcesComponent.init(Survey, domService, apollo, dialog, formBuilder);
+    OwnerComponent.init(Survey, domService, apollo);
+    UsersComponent.init(Survey, domService, apollo);
+    // load global properties
+    ReferenceDataProperties.init(Survey, domService, referenceDataService);
+  }
+
   TooltipProperty.init(Survey);
   OtherProperties.init(Survey, environment);
   // set localization
